Move Drag event into seperate file
